+++ src/index.ts
@@ -32,7 +32,13 @@ import type {
   ResultRecord,
 } from '@/src/types/result';
 import { compileQueryInput } from '@/src/utils';
-import { getProperty, omit, setProperty, splitQuery } from '@/src/utils/helpers';
+import {
+  deleteNestedProperty,
+  getProperty,
+  omit,
+  setProperty,
+  splitQuery,
+} from '@/src/utils/helpers';
 import { generatePaginationCursor } from '@/src/utils/pagination';
 
 interface TransactionOptions {
@@ -409,46 +415,10 @@ class Transaction {
 
     const fieldsToDrop = selectedFields.filter((field) => field.excluded === true);
 
-    // Helper function to delete nested property from slugs with dot notation, like
-    // `ronin.createdAt`.
-    const deleteNestedProperty = (obj: RecordType, path: string): void => {
-      const parts = path.split('.');
-      const lastPart = parts.pop()!;
-      let current = obj;
-
-      for (const part of parts) {
-        if (!current || typeof current !== 'object') return;
-        const currentAsRecord = current as Record<string, unknown>;
-        if (!(part in currentAsRecord)) return;
-        // @ts-expect-error - This is a valid field.
-        current = currentAsRecord[part];
-      }
-
-      if (typeof current === 'object' && current !== null) {
-        delete (current as Record<string, unknown>)[lastPart];
-      }
-
-      if (
-        parts.length > 0 &&
-        typeof current === 'object' &&
-        current !== null &&
-        Object.keys(current as object).length === 0
-      ) {
-        let temp = obj as Record<string, unknown>;
-        for (let i = 0; i < parts.length - 1; i++) {
-          temp = temp[parts[i]] as Record<string, unknown>;
-        }
-        const lastPart = parts.at(-1);
-        if (lastPart) {
-          delete temp[lastPart];
-        }
-      }
-    };
-
     if (fieldsToDrop.length > 0) {
       for (const record of result.records) {
         for (const field of fieldsToDrop) {
-          deleteNestedProperty(record, field.slug);
+          deleteNestedProperty(record as Record<string, unknown>, field.slug);
         }
       }
     }
